backend/requirements.txt
azure-functions==1.19.0
azure-cosmos==4.7.0
openai==1.40.0
aiohttp==3.9.5
pydantic==2.8.2
python-dotenv==1.0.1
